requirements.txt
numpy==1.22.4
opencv_python==4.6.0.66
Pillow==9.1.1
